// src/types/reward.ts
export type Reward = {
  id: string;
  name: string;
  amount: number;
  description: string | null;
  image: string;
  // needed_points: number;
  // needed_visits: number;
  // is_active: boolean;
  // is_redeemable: boolean | null;
  // limited: boolean;
  // availability: number;
  // redeem_description: string;
  // redeem_success_alert_text: string;
  // bounty_redeem_alert_header: string;
  // bounty_redeem_alert_text: string;
  // show_progress_bar: boolean;
  // shipping: boolean;
  // instant: boolean;
  // manual_claim: boolean;
  // is_expired: boolean;
  // order: number | null;
  // reward_type: number;
  // type: number;
  // app_form: any | null;
  // promo_redeem_types: Array<{ type: string }>;
  // show_confirmation_dialog: string | null;
  // valid_from: string | null;
  // valid_until: string | null;
  // pictures: any[];
  // variations: any | null;
  // prizes: any | null;
  // terms: any | null;
  // category: any | null;
  // shop: any | null;
  // condition_id: any | null;
  // point_pool: any | null;
  // adjust_event_tokens: any | null;
  // is_free_reward_usage_enabled: boolean;
  // contest: boolean;
  // highscore_contest: boolean;
  // point_contest: boolean;
  // show_ranking: boolean;
  // contest_type: any | null;
  // ranks: number;
  // cr_rank: number;
  // cr_points: any | null;
  // throttle: any | null;
  // time_range_redeem_count: any | null;
  // redeem_count: any | null;
  // single_use: boolean;
  // require_shop_when_redeemed: boolean;
  // consume_points: boolean;
  // can_participate: boolean | null;
  // is_participating: boolean | null;
  // required_status: any | null;
  // is_activateable: boolean | null;
  // is_deactivatable: boolean | null;
  // is_activated: boolean | null;
  // activation_description: string;
  // bounty_activate_alert_header: string;
  // bounty_activate_alert_text: string;
};

export type BountiesResponse = {
  count: number;
  next: string | null;
  previous: string | null;
  results: Reward[];
};